feat: 重构了代码结构 解耦合 添加了交易记录处理
